requirements.txt
# Core dependencies for RAG-based AI system
pandas>=1.5.0,<2.0.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.1.0,<2.0.0
joblib>=1.2.0
requests>=2.28.0
openai-whisper>=20231117

# OpenAI API integration
openai>=1.0.0
python-dotenv>=1.0.0

# Optional: For better performance
# torch>=1.13.0  # Uncomment if you want to use GPU acceleration for Whisper
# torchaudio>=0.13.0  # Uncomment if you need audio processing features

# Development dependencies (optional)
# jupyter>=1.0.0
# matplotlib>=3.5.0
# seaborn>=0.11.0
